% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntCFest.R
\name{syntCFest}
\alias{syntCFest}
\title{Estimate the effect of the policy or program using DID}
\usage{
syntCFest(SCFobj, use.weights = F, ...)
}
\arguments{
\item{SCFobj}{Object of class syntCFmetrics.}

\item{use.weights}{Logical. if TRUE use the quantile interval prediction ratio as weigths in the final regression model.}

\item{...}{Other arguments to be passed to the jtools and lm().}
}
\value{
Printed output for the final regression model.
}
\description{
Estimate the effect of the treatment using difference-in-difference (DID) to account for any systematic error in the trained model.
The testing period is used in combination
}
\examples{

\dontrun{
estEffect<-syntCFest(modelEval)
estEffect$Estimates
}
}
