% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntCFmetrics.R
\name{syntCFmetrics}
\alias{syntCFmetrics}
\title{Compute evaluation metrics to measure regression performance.}
\usage{
syntCFmetrics(SCFobj, quantiles = c(0.1, 0.9), ...)
}
\arguments{
\item{SCFobj}{Object of class syntCF.}

\item{quantiles}{Vector of quantiles for quantile prediction.}

\item{...}{Other arguments to be passed to predict.}
}
\value{
an object of class syntCF with the values for RMSPE, MAPE,R2 and the train and test set and vector of scoring rules metrics.
}
\description{
RMSE,MAPE and R2, are accompanied with additional metrics specifically designed to evaluate the goodness of the prediction intervals including
interval score,sharpness, underprediction, overprediction.
}
\examples{

\dontrun{
modelEval<-syntCFmetrics(trainModel,quantiles=c(0.1, 0.9))
modelEval$trainr$trainMetrics
modelEval$testr$testMetrics
modelEval$quants$intervalMetrics
}

}
