% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntCFplot.R
\name{syntCFplot}
\alias{syntCFplot}
\title{Printing and plotting a syntCFest object
Methods for printing and plotting the results of an analysis results object returned by syntCFest().}
\usage{
syntCFplot(SCFobj, ...)
}
\arguments{
\item{SCFobj}{Object of class syntCFest.}

\item{...}{Other arguments to be passed to the ggplot2}
}
\value{
returns a ggplot2 plot.
}
\description{
Printing and plotting a syntCFest object
Methods for printing and plotting the results of an analysis results object returned by syntCFest().
}
\examples{

\dontrun{
syntCFplot(estEffect)
}
}
